# swagger.yaml
openapi: 3.0.3
info:
  title: Haruki Database Backend API
  description: |
    Haruki Database Backend API 文档
    
    ## 认证
    大多数 API 需要在请求头中添加 `Authorization` 字段进行认证。
  version: 2.0.0
  contact:
    name: Haruki Dev Team
servers:
  - url: http://localhost:9621
    description: 本地开发环境
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: API 密钥认证
...
paths:
...
  /censor/name:
    post:
      tags:
        - Censor
      summary: 审核名称
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - haruki_user_id
                - server
                - user_id
                - name
              properties:
                haruki_user_id:
                  type: integer
                  description: Haruki 用户 ID
                server:
                  type: string
                  description: 服务器
                user_id:
                  type: string
                  description: 游戏用户 ID
                name:
                  type: string
                  description: 待审核名称
      responses:
        '200':
          description: 审核结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    enum: [compliant, non_compliant]
        '400':
          description: 请求参数错误
  /censor/short-bio:
    post:
      tags:
        - Censor
      summary: 审核简介
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - haruki_user_id
                - server
                - user_id
                - content
              properties:
                haruki_user_id:
                  type: integer
                  description: Haruki 用户 ID
                server:
                  type: string
                  description: 服务器
                user_id:
                  type: string
                  description: 游戏用户 ID
                content:
                  type: string
                  description: 待审核简介内容
      responses:
        '200':
          description: 审核结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    enum: [compliant, non_compliant]
        '400':
          description: 请求参数错误